feat(Avatar): Add src prop to avatar
This allows the src prop to be set on the avatar component to control
the image that is used for the avatar. It also moves the Avatar out to
be a root level component outside the card context.
