+++ src/components/Card/index.js
@@ -7,4 +7,3 @@ export { default as Primary } from './Primary';
 export { default as Subtitle } from './Subtitle';
 export { default as SupportingText } from './SupportingText';
 export { default as Title } from './Title';
-export { default as Avatar } from './Avatar';
